# docs/swagger.yaml
definitions:
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
...
  reqmodel.RatedMovieListResponse:
    properties:
      rated_movie_list:
        items:
          $ref: '#/definitions/sqlc.GetMovieRatingListRow'
        type: array
      user_id:
        type: string
    type: object
  reqmodel.RatedMovieRequest:
    properties:
      movie_id:
        type: string
      rating:
        type: integer
    type: object
  reqmodel.UserListResponse:
    properties:
      user_list:
        items:
          $ref: '#/definitions/sqlc.UserDatum'
        type: array
    type: object
  reqmodel.UserRequest:
    properties:
      login:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  sqlc.GetMovieRatingListRow:
    properties:
      movie_id:
        type: string
      rating:
        type: integer
    type: object
...
  sqlc.RatedMovie:
    properties:
      movie_id:
        type: string
      rating:
        type: integer
      user_id:
        type: string
    type: object
  sqlc.UserDatum:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: string
      login:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Basic swagger for current api
  termsOfService: http://swagger.io/terms/
  title: movie_backend_go
  version: "1.0"
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Get user list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.UserListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Show user list
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - description: User creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sqlc.UserDatum'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create user
      tags:
      - user
  /user/{user_id}:
    delete:
      consumes:
      - application/json
      description: Delete user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get user by id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.UserDatum'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Show user
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: User creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.UserDatum'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user
      tags:
      - user
...
  /user/{user_id}/rated_movie:
    get:
      consumes:
      - application/json
      description: Get movie rating by user id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.RatedMovieListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get  movie rating
      tags:
      - rated_movie
    patch:
      consumes:
      - application/json
      description: Update movie rating
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Updated rating
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.RatedMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.RatedMovie'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update movie rating
      tags:
      - rated_movie
    post:
      consumes:
      - application/json
      description: Rate movie
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Rate movie data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.RatedMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.RatedMovie'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get  movie rating
      tags:
      - rated_movie
  /user/{user_id}/rated_movie/{movie_id}:
    delete:
      consumes:
      - application/json
      description: Delete certain movie rating
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete movie rating
      tags:
      - rated_movie
swagger: "2.0"